Misc changes
Added short description to http links, images and videos.
Renamed "published_work.md" file to "scientific_publications.
In README.md, a copy of "Linkedin" logo was made and added an argument to decrease the size by 50%. Although it is not compatible with GitHub pages, the code was kept for possible futur implementation.
A link to the .pdf files was added to each paper in the "scientific_publications.md" file. Also, the gaps in the file name were replaced by an underscore to be compatible with Markdown.